Autherization and Authentication
There is now an admin account seeded into the SQL server with special pages only accessables to the user or admin accordingly. The pages could use visual update from the baseline asp theme... But all CRUD operations are covered. Next we add a few things to the home page and maybe make it prettier.